chore: add simple check before schema files copy

+++ mailbox-sdk/copy-soap-schemas.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [[ ! -e '../soap/target/classes/com/zimbra/soap/ZimbraService.wsdl' ]]; then
+  echo "Missing soap artifacts."
+  echo "Please run 'mvn package' in the mailbox's soap module."
+  exit 1
+fi
+
 rm -rf schemas
 mkdir schemas
 cp ../soap/target/classes/com/zimbra/soap/ZimbraService.wsdl schemas/ZimbraService.wsdl
